fix(mypy): _heldout_smooth_r2_fast returns Any through the untyped prep tuple
prep is annotated as a bare tuple, so unpacking sst from it gives Any, and
1.0 - sse / sst silently returned Any from a function declared -> float.
Wrap the final division in float() to match the declared return type.

Also adds `concepts` to deptry's DEP003 ignore list: it has the identical
DEP001-vs-DEP003 classification flip already documented for `finance` (a
locally-installed optional dep that deptry reports differently depending on
whether it happens to be resolvable in the environment deptry runs in), which
started blocking this commit's pre-commit deptry gate.

+++ src/mlframe/feature_selection/filters/_orthogonal_univariate_fe/_orth_extra_basis_fe.py
@@ -829,7 +829,7 @@ def _heldout_smooth_r2_fast(x: np.ndarray, prep: tuple) -> float:
         return 0.0
     pred = np.vander(zx[va], 4) @ coef
     sse = float(np.sum((yv - pred) ** 2))
-    return 1.0 - sse / sst
+    return float(1.0 - sse / sst)
 
 
 def _detect_fourier_freq_for_col(
